requirements.txt
# 核心依赖
openai-whisper>=20231117
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0

# 下载工具
yt-dlp>=2023.11.16

# 音频处理
ffmpeg-python>=0.2.0

# 网页解析
requests>=2.31.0
beautifulsoup4>=4.12.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 配置文件
pyyaml>=6.0

# 其他
tqdm>=4.66.0